[TASK] Fix phpstan checkFunctionArgumentTypes errors in ext:core Crypto
This patch fixes incompatible type usage in function arguments
and is preparatory work for introducing native type hints and
strict mode in all core files.

Releases: master, 10.4
Resolves: #92262
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/66038
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>

+++ typo3/sysext/core/Classes/Crypto/PasswordHashing/PhpassPasswordHash.php
@@ -252,7 +252,7 @@ class PhpassPasswordHash implements PasswordHashInterface
             if (!strncmp('$', $salt, 1)) {
                 if (!strncmp(self::PREFIX, $salt, strlen(self::PREFIX))) {
                     $isValid = true;
-                    $salt = substr($salt, strrpos($salt, '$') + 2);
+                    $salt = substr($salt, (int)strrpos($salt, '$') + 2);
                 } else {
                     $skip = true;
                 }
